feat: shared module
