+++ frontend/src/app/app.component.html
@@ -13,3 +13,4 @@
     </div>
   </div>
 </div>
+<router-outlet></router-outlet>
