+++ frontend/src/app/application.constant.ts
@@ -1,0 +1,2 @@
+export class ApplicationConstant {
+}
