+++ frontend/src/app/modules/core/module-import-guard.ts
@@ -1,0 +1,5 @@
+export function throwIfAlreadyLoaded(parentModule: any, moduleName: string): void {
+  if (parentModule) {
+    throw new Error(`${moduleName} has already been loaded. Import ${moduleName} in the AppModule only.`);
+  }
+}
